requirements.txt
streamlit
python-dotenv
groq
pillow